Added support to HELO and QUIT commands and other changes
- Support to HELO and QUIT commands has been added

Change has been made on SMTPMessage class:
- On the main constructor, it will automactically identify the message
received and set the associated command id to mCommandId attribute and
will save the content of the received message into the mArguments
attribute. Note that the command will not be saved into the mArguments
attribute.
Example:
if "HELO hi" is received, the content of mArgument will be "hi"